pyproject.toml
[tool.poetry]
name = "service-sector"
version = "0.1.0"
description = ""
authors = ["Yuriy Kormin <yuriy.kormin@gmail.com>"]
readme = "README.md"
packages = [{include = "app/territory_sectors"}]

[tool.poetry.dependencies]
#python = "^3.10"
python = ">=3.10,<3.12"
python-dotenv = "^0.21.0"
dj-database-url = "^1.2.0"
django-bootstrap4 = "^22.3"
postgis = "^1.0.4"
shortuuid = "^1.0.11"
psycopg2-binary = "^2.9.5"
django-debug-toolbar = "^3.8.1"
numpy = "^1.24.1"
gunicorn = "^20.1.0"
django-qr-code = "^3.1.1"
whitenoise = "^6.3.0"
pillow = "^9.4.0"
qrcode = "^7.3.1"
django-simple-history = "^3.2.0"
rollbar = "^0.16.3"
coverage = "^7.2.1"
django-filter = "^22.1"
reportlab = "^3.6.12"
geojson = "^3.0.1"
natsort = "^8.3.1"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
